Feature #2379 main_v11.1 sonarqube updates (#2851)

+++ internal/scripts/docker/build_met_sonarqube.sh
@@ -5,7 +5,7 @@
 #
 # This build_met_sonarqube.sh script must be run from the top-level
 # directory of the MET repository to be analyzed. It runs SonarQube to
-# scan both the Python and C/C++ MET source code.
+# scan the MET source code.
 #
 # Usage: internal/scripts/docker/build_met_sonarqube.sh
 #
@@ -75,46 +75,29 @@ if [ -z ${SONARQUBE_OUT_DIR} ]; then
   export SONARQUBE_OUT_DIR=bw-outputs
 fi
 
+# Define the version string
+SONAR_PROJECT_VERSION=$(cat docs/version | cut -d'=' -f2 | tr -d '" ')
+
 # Store the full path to the scripts directory
 SONAR_PROPERTIES_DIR=internal/scripts/sonarqube
 SONAR_PROPERTIES=sonar-project.properties
 
-# Copy sonar-project.properties for Python code
+# Configure the sonar-project.properties
 [ -e $SONAR_PROPERTIES ] && rm $SONAR_PROPERTIES
-sed -e "s|SONAR_TOKEN|$SONAR_TOKEN|" \
+sed -e "s|SONAR_PROJECT_KEY|MET-GHA|" \
+    -e "s|SONAR_PROJECT_NAME|MET GHA|" \
+    -e "s|SONAR_PROJECT_VERSION|$SONAR_PROJECT_VERSION|" \
     -e "s|SONAR_HOST_URL|$SONAR_HOST_URL|" \
-    -e "s|SONAR_PROJECT_KEY|MET-GHA-Python|" \
-    -e "s|SONAR_PROJECT_NAME|MET GHA Python|" \
+    -e "s|SONAR_TOKEN|$SONAR_TOKEN|" \
     -e "s|SONAR_BRANCH_NAME|$MET_GIT_NAME|" \
-    $SONAR_PROPERTIES_DIR/python.sonar-project.properties > $SONAR_PROPERTIES
+    $SONAR_PROPERTIES_DIR/$SONAR_PROPERTIES > $SONAR_PROPERTIES
 
 # The source and reference branches must differ to define new code
 if [ "$MET_GIT_NAME" != "$SONAR_REFERENCE_BRANCH" ]; then
   echo "sonar.newCode.referenceBranch=${SONAR_REFERENCE_BRANCH}" >> $SONAR_PROPERTIES
 fi
 
-# Run SonarQube scan for Python code
-time_command $SONAR_SCANNER
-
-# Copy the Python scan report-task.txt file
-mkdir -p /met/.scannerwork
-cp .scannerwork/report-task.txt /met/.scannerwork/python-report-task.txt
-
-# Copy sonar-project.properties for C/C++ code
-[ -e $SONAR_PROPERTIES ] && rm $SONAR_PROPERTIES
-sed -e "s|SONAR_TOKEN|$SONAR_TOKEN|" \
-    -e "s|SONAR_HOST_URL|$SONAR_HOST_URL|" \
-    -e "s|SONAR_PROJECT_KEY|MET-GHA-CXX|" \
-    -e "s|SONAR_PROJECT_NAME|MET GHA CXX|" \
-    -e "s|SONAR_BRANCH_NAME|$MET_GIT_NAME|" \
-    $SONAR_PROPERTIES_DIR/sonar-project.properties > $SONAR_PROPERTIES
-
-# The source and reference branches must differ to define new code
-if [ "$MET_GIT_NAME" != "$SONAR_REFERENCE_BRANCH" ]; then
-  echo "sonar.newCode.referenceBranch=${SONAR_REFERENCE_BRANCH}" >> $SONAR_PROPERTIES
-fi
-
-# Run the configure script
+# Run the MET configure script
 time_command ./configure \
   BUFRLIB_NAME=${BUFRLIB_NAME} \
   GRIB2CLIB_NAME=${GRIB2CLIB_NAME} \
@@ -122,7 +105,7 @@ time_command ./configure \
   --enable-modis \
   --enable-mode_graphics \
   --enable-lidar2nc \
-  --enable-python
+  --enable-python \
   CPPFLAGS="-I/usr/local/include -I/usr/local/include/freetype2 -I/usr/local/include/cairo" \
   LIBS="-ltirpc"
 
@@ -132,11 +115,18 @@ time_command make clean
 # Run SonarQube make
 time_command $SONAR_WRAPPER --out-dir $SONARQUBE_OUT_DIR make
 
-# Run SonarQube scan for C/C++ code
+# Run SonarQube scan
 time_command $SONAR_SCANNER
+status=$?
 
-# Copy the C/C++ scan report-task.txt file
+# Check return status
+if [[ $status -ne 0 ]]; then
+  echo "ERROR: ${0} -> the SonarQube scan returned with non-zero status (${status})!"
+  exit ${status}
+fi
+
+# Copy the scan report-task.txt file
 mkdir -p /met/.scannerwork
-cp .scannerwork/report-task.txt /met/.scannerwork/cxx-report-task.txt
+cp .scannerwork/report-task.txt /met/.scannerwork/report-task.txt
 
 [ -e $SONAR_PROPERTIES ] && rm $SONAR_PROPERTIES
